Разгрузил компоненты отрисовки событий. Перенес бизнес-логику в Trip Presenter
