+++ src/view/events-plan-container.js
@@ -1,219 +1,10 @@
-import {render, AddedComponentPosition} from "../utils/render.js";
-import TripEventsItemView from "./trip-events-item.js";
-import TripDaysItemView from "./trip-days-item.js";
-import EventsListView from "./events-list.js";
-import {cities, eventTypes, getOffers} from "../mock/event.js";
-import BaseEventView from "./base-event.js";
-import SelectedOffersContainerView from "./selected-offers-container.js";
-import OpenEventButtonView from "./open-event-button.js";
-import ReadingEventContentView from "./reading-event-content.js";
-import OfferItemView from "./offer-item.js";
 import AbstractView from "./abstract.js";
-
-const EMPTY_EVENT_INDEX = 0;
-const MAX_OFFERS_COUNT = 3;
 
 const createEventsPlanContainerTemplate = () =>
   `<ul class="trip-days"></ul>`;
 
 export default class EventsPlanContainer extends AbstractView {
-  constructor(events) {
-    super();
-    this._events = events;
-  }
-
   getTemplate() {
     return createEventsPlanContainerTemplate();
-  }
-
-  // Формирование структуры событий по датам
-  _getMapDates() {
-    const datesSet = new Set();
-
-    this._events.forEach((evt) => {
-      const date = new Date(evt.timeInterval.leftLimitDate);
-      date.setHours(0, 0, 0, 0);
-
-      datesSet.add(date.getTime());
-    });
-
-    const dates = Array.from(datesSet)
-      .sort((a, b) => a - b);
-    const mapDates = new Map();
-    dates.forEach((date) => mapDates.set(date, []));
-
-    this._events.forEach((evt) => {
-      const date = new Date(evt.timeInterval.leftLimitDate);
-      date.setHours(0, 0, 0, 0);
-
-      mapDates.get(date.getTime()).push(evt);
-    });
-
-    return mapDates;
-  }
-
-  // Значения по умолчанию для события при создании события
-  _getDefaultEvent() {
-    const tmpCities = Array.from(cities.keys());
-    const evt = {
-      eventType: eventTypes[EMPTY_EVENT_INDEX],
-      city: tmpCities[EMPTY_EVENT_INDEX],
-      offers: [],
-      destination: cities.get(tmpCities[EMPTY_EVENT_INDEX]),
-      isFavorite: false,
-      price: 0
-    };
-
-    const date = new Date();
-    date.setHours(0, 0, 0);
-
-    evt.timeInterval = {
-      leftLimitDate: date,
-      rightLimitDate: date
-    };
-
-    const offers = getOffers(evt.eventType);
-    for (let i = 0; i < offers.length; i++) {
-      evt.offers[i] = {
-        name: offers[i].name,
-        price: offers[i].price,
-        isAccepted: false
-      };
-    }
-
-    return evt;
-  }
-
-  // Отрисовка события и его подмена на форму редактирования
-  _renderEvent(evt, tripEventsItemView) {
-    const isNewEvent = evt === null;
-    evt = evt || this._getDefaultEvent();
-
-    // Содержимое события (для чтения)
-    const readingEventContentView = new ReadingEventContentView(evt);
-
-    // Форма добавление/редактирования
-    const eventView = new BaseEventView(evt, isNewEvent);
-    eventView.fillEvent();
-
-    // Подмена события на форму редактирования
-    const replaceEventToForm = () => {
-      tripEventsItemView.replace(
-          eventView,
-          readingEventContentView
-      );
-    };
-
-    // Подмена формы редактирования на событие
-    const replaceFormToEvent = () => {
-      tripEventsItemView.replace(
-          readingEventContentView,
-          eventView
-      );
-    };
-
-    const onEscKeyDown = (evt1) => {
-      if (evt1.key === `Escape` || evt1.key === `Esc`) {
-        evt1.preventDefault();
-        replaceFormToEvent();
-        document.removeEventListener(`keydown`, onEscKeyDown);
-      }
-    };
-
-    // Отрисовка события
-    render(
-        tripEventsItemView,
-        readingEventContentView,
-        AddedComponentPosition.BEFORE_END
-    );
-
-    // Контейнер предложений для текущего события
-    const selectedOffersContainer = new SelectedOffersContainerView();
-    render(
-        readingEventContentView,
-        selectedOffersContainer,
-        AddedComponentPosition.BEFORE_END
-    );
-
-    // Заполнение контейнера предложений текущего события
-    let cnt = 0; // счетчик предложений для короткой записи события
-    for (let k = 0; k < evt.offers.length; ++k) {
-      if (evt.offers[k].isAccepted) {
-        ++cnt;
-        if (cnt > MAX_OFFERS_COUNT) {
-          break;
-        }
-
-        render(
-            selectedOffersContainer,
-            new OfferItemView(evt.offers[k]),
-            AddedComponentPosition.BEFORE_END
-        );
-      }
-    }
-
-    // В конце элемента для чтения кнопка открытия события
-    render(
-        readingEventContentView,
-        new OpenEventButtonView(),
-        AddedComponentPosition.BEFORE_END
-    );
-
-    // Подписка на кнопку открытия формы
-    readingEventContentView.setEditClickHandler(() => {
-      replaceEventToForm();
-      document.addEventListener(`keydown`, onEscKeyDown);
-    });
-
-    // Подписка на отправку формы
-    eventView.setFormSubmitHandler(() => {
-      replaceFormToEvent();
-      document.removeEventListener(`keydown`, onEscKeyDown);
-    });
-  }
-
-  // Формирование дерева событий с возвратом ее структуры
-  fillTree() {
-    const mapDates = this._getMapDates();
-
-    let index = 0;
-    for (const mapDateKey of mapDates.keys()) {
-      // Какая-то дата путешествия
-      const date = new Date(mapDateKey);
-
-      // Отрисовка очередной даты
-      const tripDaysItemView = new TripDaysItemView(date, index++);
-      render(
-          this,
-          tripDaysItemView,
-          AddedComponentPosition.BEFORE_END
-      );
-
-      const eventsListView = new EventsListView();
-      render(
-          tripDaysItemView,
-          eventsListView,
-          AddedComponentPosition.BEFORE_END
-      );
-
-      // события даты сортируем по дате начала
-      const tmpEvents = mapDates.get(mapDateKey)
-        .sort((a, b) => a.timeInterval.leftLimitDate.getTime() -
-                        b.timeInterval.leftLimitDate.getTime());
-      // Цикл по всем событиям данной даты
-      for (let j = 0; j < tmpEvents.length; ++j) {
-        // Контейнер события
-        const tripEventsItemView = new TripEventsItemView(tmpEvents[j]);
-        render(
-            eventsListView,
-            tripEventsItemView,
-            AddedComponentPosition.BEFORE_END
-        );
-
-        this._renderEvent(tmpEvents[j], tripEventsItemView);
-      }
-    }
-
-    return mapDates;
   }
 }
